Basic refactoring of ChEMBL client
